requirements.txt
astroid==2.4.2
autopep8==1.5.4
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.3
cryptography==3.4.8
cycler==0.10.0
fonttools==4.25.0
idna==3.2
isort==5.5.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.4.3
mccabe==0.6.1
mkl-service==2.3.0
munkres==1.1.4
numexpr==2.7.3
numpy==1.21.3
oauthlib==3.1.1
olefile==0.46
pandas==1.3.3
Pillow==8.4.0
pip==21.0.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.6.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-qt5==0.1.10
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
setuptools==58.0.4
sip==4.19.13
six==1.15.0
toml==0.10.1
torch==1.6.0
tornado==6.1
tweepy==4.1.0
urllib3==1.26.7
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
